[experimental] allow compilation with stable Rust
Cf. simoncozens/ufostroker#1.

No one really uses `PointData` yet, so this is quite safe; even
internally it's unused. With it being `impl`'d on (), the most common T
for glifparser::Point<T>, this _should_ be fine.

If you use it all it means is you need another line of boilerplate…

```rust
impl PointData for MyDataType {}
```